mcad/requirements.txt
solidpython>=1.1.2
black>=21.11b1
regex>=2021.4.4
pylint>=2.12.2
mypy>=0.910
